pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pyyaml>=6.0.2",
    "streamlit-authenticator>=0.4.2",
    "streamlit>=1.45.0",
    "bcrypt>=4.3.0",
]
